agent/requirements.txt
pywin32==306
psutil==5.9.8
pystray==0.19.5
Pillow==10.2.0
requests==2.31.0
pynput==1.7.6
schedule==1.2.1
screeninfo==0.8.1
watchdog==4.0.0
pyinstaller==6.3.0
# WebRTC live screen streaming
aiortc==1.9.0
av==12.3.0
numpy==1.26.4
websockets==12.0
